requirements.txt
python-telegram-bot==20.7
python-pptx==0.6.23
openai==1.12.0
aiohttp==3.9.3
